First commit for METIS partitioning
This is going to be quite complicated... all the code needs to be refactored so that instead of iterating over %R, %L, %F, %B etc, we maintain information about which nodes/beams are shared and pass between partitions as required.

The actual partitioning of the domain was really easy, just an API call to METIS - this part is done.