Change computation of melange pressure
Previously, was summing frx,fry,frz per particle, then finding
magnitude of the result, but this is an underestimate, which differs
from the result if summing the force magnitudes directly, because
forces aren't properly aligned. The new way is (I believe?!) the
accurate representation of 'pressure' in the particle.

Also re-added the removal of outliers (too close particles)