Add some rotation/normal vector stuff to python/README
I am not 100% sure that the rotation -> normal vector stuff is
correct. I think it is correct for 'HiDEM' coordinates - i.e. assuming
that the particles begin facing XYZ. However, if we're rotating stuff
to 'real world' coords, I think we might need to add some initial
offset to the rotation (because rotation is relative to the initial position)